+++ src/main/java/com/objects/marketbridge/common/infra/KakaoPayService.java
@@ -68,7 +68,7 @@ public class KakaoPayService {
     }
 
 
-    //인터페이스 -> 정기구독 정기용, 단건용
+    //정기구독 1회차 , 단건결제
     public KakaoPayApproveResponse approve(KakaoPayApproveRequest request) {
 
         MultiValueMap<String, String> requestMap = request.toMultiValueMap();
@@ -82,6 +82,7 @@ public class KakaoPayService {
                 .body(KakaoPayApproveResponse.class);
     }
 
+    //정기구독 2회차
     public KakaoPaySubsApproveResponse subsApprove(KakaoPaySubsApproveRequest request) {
 
         MultiValueMap<String, String> requestMap = request.toMultiValueMap();
@@ -143,7 +144,7 @@ public class KakaoPayService {
                 .messageConverters((converters) ->
                         converters.add(new FormHttpMessageConverter()))
                 .defaultHeaders((httpHeaders -> {
-                    httpHeaders.add(AUTHORIZATION, KAKAO_AK + kakaoPayConfig.getAdminKey());
+                    httpHeaders.add(AUTHORIZATION, kakaoPayConfig.getSecretKeyDev());
                     httpHeaders.add(ACCEPT, APPLICATION_JSON.toString());
                     httpHeaders.add(CONTENT_TYPE, APPLICATION_FORM_URLENCODED + ";charset=UTF-8");
                 }))
