+++ src/main/java/com/objects/marketbridge/member/service/MemberService.java
@@ -24,7 +24,6 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class MemberService {
 
-    private final MembershipRepository membershipRepository;
     private final MemberRepository memberRepository;
 
     public CheckedResultDto isDuplicateEmail(String email){
@@ -32,8 +31,8 @@ public class MemberService {
         return CheckedResultDto.builder().checked(isDuplicateEmail).build();
     }
 
-    public void addMemberAddress(){
-
+    public void addMemberAddress(Long id){
+        memberRepository.findByIdWithAddresses(id);
     }
 
 
@@ -42,78 +41,4 @@ public class MemberService {
         memberRepository.save(member);
     }
 
-    @Transactional
-    public void changeMemberShip(Long id){
-
-        // 이제 orElseThrow 는 전부 MemberRepositoryImpl 에서 처리하기로 했습니다. by 정민우
-//        Member findMember = memberRepository.findById(id)
-//                .orElseThrow(() -> new IllegalArgumentException("Member not found with id: " + id)); // id 를 통한 조회실패 예외발생
-
-        Member findMember = memberRepository.findById(id);
-
-        if(findMember.getMembership().equals("BASIC")){//멤버십 WOW 등록
-            findMember.setMembership(MembershipType.WOW.toString());
-            memberRepository.save(findMember);
-        }else {// 멤버십 BASIC으로 해제
-            findMember.setMembership(MembershipType.BASIC.toString());
-            memberRepository.save(findMember);
-        }
-    }
-
-    @Transactional
-    public void savePayReadyData(CreateSubsDto createSubsDto) { //CreateorderService
-
-        // 1. sid 정보 저장
-        Membership membership = membershipRepository.save(createSubscribtion(createSubsDto));
-
-    }
-
-    @Transactional
-    public void saveAprrovalResponse(KakaoPayApproveResponse response){
-        membershipRepository.save(saveSubsResponse(response));
-    }
-
-    @Transactional
-    public void saveSubsAprrovalResponse(KakaoPaySubsApproveResponse response){
-        membershipRepository.save(saveSubscribePaymentResponse(response));
-    }
-
-    private Membership createSubscribtion(CreateSubsDto createSubsDto) {
-        Member member = memberRepository.findById(createSubsDto.getMemberId());
-        String subsOrderNo = createSubsDto.getSubsOrderNo();
-        String tid = createSubsDto.getTid();
-        Amount amount = createSubsDto.getAmout();
-
-        return Membership.create(member,subsOrderNo, tid,amount);
-    }
-
-    private Membership saveSubsResponse(KakaoPayApproveResponse response){
-        String tid = response.getTid();
-        String sid = response.getSid();
-        String cid = response.getCid();
-        String orderNo = response.getPartnerOrderId();
-        String partnerUserId = response.getPartnerUserId();
-        String itemName = response.getOrderName();
-        Long quantity = response.getQuantity();
-        String paymentMethod = response.getPaymentMethodType();
-        CardInfo cardInfo = response.getCardInfo();
-        Amount amount = response.getAmount();
-
-        return Membership.createPayment(tid, sid, cid, orderNo,partnerUserId ,itemName ,quantity, paymentMethod, cardInfo, amount);
-    }
-
-    private Membership saveSubscribePaymentResponse(KakaoPaySubsApproveResponse response){
-        String tid = response.getTid();
-        String sid = response.getSid();
-        String cid = response.getCid();
-        String orderNo = response.getPartnerOrderId();
-        String partnerUserId = response.getPartnerUserId();
-        String itemName = response.getOrderName();
-        Long quantity = response.getQuantity();
-        String paymentMethod = response.getPaymentMethodType();
-        CardInfo cardInfo = response.getCardInfo();
-        Amount amount = response.getAmount();
-
-        return Membership.createPayment(tid, sid, cid, orderNo,partnerUserId ,itemName ,quantity, paymentMethod, cardInfo, amount);
-    }
 }
