+++ src/main/resources/application.yml
@@ -51,8 +51,10 @@ logging:
 
 payment:
   kakao:
-    admin_key : "38ef7d9f7a2a2a43c93de1f5bb3ea2f5"
-    test_api_key : "13d26c106e60b48a0ae5d8280e0ceb10"
+    client_id : "AE7564ADF66B30123C2A"
+    client_secret : "94B65EF567BCB6E860B4"
+    secret_key: "PRD4A3CDADAF6C09168BA3E4725A3C678A78B271"
+    secret_key_dev : "DEVEC2152904951CB5CF3C5BEEBB3C84B6AD6F26"
 
 host: "http://localhost:8080"
 
@@ -103,7 +105,9 @@ logging:
 
 payment:
   kakao:
-    admin_key: "38ef7d9f7a2a2a43c93de1f5bb3ea2f5"
-    test_api_key: "13d26c106e60b48a0ae5d8280e0ceb10"
+    client_id : "AE7564ADF66B30123C2A"
+    client_secret : "94B65EF567BCB6E860B4"
+    secret_key: "PRD4A3CDADAF6C09168BA3E4725A3C678A78B271"
+    secret_key_dev : "DEVEC2152904951CB5CF3C5BEEBB3C84B6AD6F26"
 
 host: "http://localhost:8080/"
