+++ src/test/resources/schema-h2.sql
@@ -25,11 +25,16 @@ CREATE TABLE membership (
                             primary key,
 
                         member_id	varchar(255)	,
+                        cid varchar(255)	,
                         tid	varchar(255)	,
                         sid	varchar(255)	,
+                        item_name varchar(255),
                         subs_order_no	varchar(255)	,
+                        partner_order_id varchar(255)   ,
+                        partner_user_id varchar(255)   ,
                         status_code	varchar(255)	,
                         payment_method	varchar(255)	,
+                        quantity bigint ,
                         total_amount bigint,
                         discount_amount bigint,
                         card_issuer_name	varchar(255)	,
